Harden the rejection path: no signed overflow, plus a window-edge test
From the adversarial review: in_window computed price minus base_ before the range check, which is signed overflow UB for extreme prices on the path the class advertises as safe rejection. Compute the offset with unsigned arithmetic in both in_window and index_of. Add a unit test that rejects prices just outside both window edges and an extreme price, asserting the book is unchanged, plus a Replace to an out-of-window price that leaves the original order intact. Verified under UBSan.

+++ include/eib/order_book.hpp
@@ -137,7 +137,8 @@ class OrderBook {
   bool in_window(Ticks price, std::size_t& idx) const;
   // For prices already known to be in the window (a resting order's own price).
   std::size_t index_of(Ticks price) const {
-    return static_cast<std::size_t>(price - base_);
+    return static_cast<std::size_t>(static_cast<std::uint64_t>(price) -
+                                    static_cast<std::uint64_t>(base_));
   }
   void add_level_qty(Side side, std::size_t idx, Qty qty);
   void remove_level_qty(Side side, std::size_t idx, Qty amount);
